YAY-18 refactor: split usePostController into dedicated hooks
