+++ src/app/(private)/profile/[login]/page.tsx
@@ -2,11 +2,11 @@
 
 import { withPrivateRoute } from '@/shared/HOC'
 import { ProfileHeaderPrivate } from '@/entities/user/ui/ProfileHeader'
-import { usePostController } from '@/shared/hooks'
 import { ProfilePrivatePosts } from '@/features/posts/ui/ProfilePosts'
 import { useGetPostByIdQuery } from '@/features/posts/api'
 import { Post } from '@/entities/posts/ui/Post'
 import { EditPostModal } from '@/entities/posts/ui/modals/EditPostModal'
+import { usePostController } from '@/features/posts/hooks'
 
 function ProfilePrivatePage() {
    const {
