+++ src/features/posts/hooks/index.ts
@@ -1,1 +1,4 @@
 export * from './usePublishPost'
+export * from './usePostUrlState'
+export * from './usePostDelete'
+export * from './usePostController'
