+++ src/shared/hooks/index.ts
@@ -3,4 +3,3 @@ export * from './useAppDispatch'
 export * from './useRelativeTime'
 export * from './useAuthSync'
 export * from './useInfiniteScroll'
-export * from './usePostController'
